+++ src/main/resources/db/migration/V9__create_pagos_table.sql
@@ -2,7 +2,7 @@
 CREATE TABLE pagos (
     id UUID NOT NULL PRIMARY KEY,
     suscripcion_id UUID NOT NULL,
-    monto DOUBLE NOT NULL,
+    monto DOUBLE PRECISION NOT NULL,
     estado VARCHAR(50) NOT NULL,
     metodo_pago VARCHAR(100) NOT NULL,
     fecha_pago TIMESTAMP,
